% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_generation.R
\name{construct_weighted_graph_from_edge_list}
\alias{construct_weighted_graph_from_edge_list}
\title{Contruct weighted graph using node weights}
\usage{
construct_weighted_graph_from_edge_list(edge_list_, expr_matrix_, grouping_vec_)
}
\arguments{
\item{edge_list_}{A \code{data.frame}. First two columns will be parsed as a edge list.}
}
\description{
Contruct weighted graph using node weights
}

