% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fpkm}
\alias{fpkm}
\title{FPKM table of distal mouse lung epithelial cells from two developmental stages (E14.5 and E16.5)}
\format{A matrix with rows being the genes sequenced and columns being the samples.}
\usage{
fpkm
}
\description{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52583
}
\keyword{datasets}

