% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_generation.R
\name{generate_spinglass_communities}
\alias{generate_spinglass_communities}
\title{Generate communities using Spin Glass algorithm}
\usage{
generate_spinglass_communities(weighted_graph_, seed_genes_, n_threads_ = 1,
  gamma_ = 0.5, min_size_ = 1, max_size_ = 100, verbose_level_ = 1,
  seed_ = 12345)
}
\description{
Generate communities using Spin Glass algorithm
}

