% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_generation.R
\name{get_monte_carlo_simulation_p_values}
\alias{get_monte_carlo_simulation_p_values}
\title{Do Monte Carlo simulation}
\usage{
get_monte_carlo_simulation_p_values(module_graphs_, n_monte_carlo_ = 1000,
  n_threads_ = 1, verbose_level_ = 1, seed_ = 12345)
}
\arguments{
\item{gene_weights_}{A numeric vector. Should be the same one the user supplied to \code{construct_weighted_graph}.}
}
\description{
Do Monte Carlo simulation
}

