% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_generation.R
\name{GO_analysis}
\alias{GO_analysis}
\title{Gene Ontology (GO) term analysis}
\usage{
GO_analysis(selected_genes_, all_genes_, organism_, gene_id_type_ = "symbol",
  ontology_ = "CC", verbose_level_ = 1, pvalueCutoff_ = 0.01,
  qvalueCutoff_ = 0.05)
}
\arguments{
\item{selected_genes_}{A character vector of genes, by default gene symbols. The ID type can be changed
using \code{gene_id_type_} parameter.}

\item{all_genes_}{A character vector of genes treated as background. It should be a superset of \code{selected_genes_}.}

\item{organism_}{Either "human" or "mouse".}

\item{gene_id_type_}{Currently either "symbol" (default), or "entrez_id".}

\item{ontology_}{Either "MF" (Molecular Function), "CC" (Cellular Component, default), or "BP" (Biological Process).}

\item{verbose_level_}{0 for quiet, 1 for simple messages, 2 for additional debug messages.}

\item{pvalueCutoff_, qvalueCutoff_}{Cutoff threshold for either p or q values.}
}
\description{
Gene Ontology (GO) term analysis
}

