% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hppi}
\alias{hppi}
\title{Human Protein Protein Interaction Network}
\format{A data frame with two columns \code{g1} and \code{g2}. Each row
  indicates a pair of genes that have putative interations.}
\usage{
hppi
}
\description{
This network is extracted from Human Integrated Protein-Protein Interaction rEference.
http://cbdm-01.zdv.uni-mainz.de/~mschaefer/hippie/index.php
}
\keyword{datasets}

