% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{lambda_update}
\alias{lambda_update}
\title{Update of lambda}
\usage{
lambda_update(Xtt, Theta, Lambda)
}
\arguments{
\item{Xtt}{matrix. Correspond to the matrix multiplication of t(X) and t.}

\item{Theta}{matrix of dimension \code{M x H}.}

\item{Lambda}{matrix of dimension \code{H x K}.}
}
\value{
Returns the updated value of Lambda
}
\description{
Update of lambda
}
\keyword{internal}
