% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leading_genes}
\alias{leading_genes}
\title{Leading genes of distal mouse lung epithelial cells from two developmental stages (E14.5 and E16.5)}
\format{A character vector of the 500 leading genes determined using D-score.}
\usage{
leading_genes
}
\description{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52583
}
\keyword{datasets}

