% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{lnfact}
\alias{lnfact}
\title{Logarithm of a factorial}
\usage{
lnfact(x)
}
\arguments{
\item{x}{an integer}
}
\value{
log(factorial(x))
}
\description{
Logarithm of a factorial
}
\keyword{internal}
