% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mppi}
\alias{mppi}
\title{Mouse Protein Protein Interaction Network}
\format{A data frame with two columns \code{g1} and \code{g2}. Each row
  indicates a pair of genes that have putative interations.}
\usage{
mppi
}
\description{
This network is extracted from the interaction database provided by BioGRID.
http://thebiogrid.org/
}
\keyword{datasets}

