% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_generation.R
\name{nmf_subpopulation}
\alias{nmf_subpopulation}
\title{Detect subpopulations using NMF}
\usage{
nmf_subpopulation(expr_matrix_, log_transformation_ = T, verbose_level_ = 1,
  n_threads_ = 1, nrun_ = 30, method_ = "brunet",
  top_genes_facet_title_font_size_ = 24, .options_ = sprintf("p\%dv\%d",
  n_threads_, verbose_level_ - 1), seed_ = 12345)
}
\arguments{
\item{expr_matrix_}{The expression matrix, each row is a gene and each
column is a sample. The row names should be gene symbols and the column
names should be sample ids. Use the correct capticalization for gene
symbols. For mouse genes,only the first letter is capitalized (e.g. Tp53);
for human genes, all letters are capitalized (e.g. TP53). The values will
be log-transformed by default, this can be changed using the
\code{log_transofrmation_} option.}

\item{log_transformation_}{Whether to log-transform the expression matrix.
A pseudo count \code{1} will be added to each value before log transformation
to avoid infinity. That is, \code{expr_matrix_ <- log(expr_matrix_+1)}. Default: \code{True}.}

\item{verbose_level_}{How much information is printed. 0 = quiet, 1 = normal, 2 = with debug info, 3 = with extra debug info. Default: 1.}

\item{n_threads_, nrun_, method_, .options_, seed_}{Parameters for \code{nmf(...)}.
Default values are supplied.}

\item{method_}{Interative method for NMF. See the documentation of \code{NMF::nmf}.}

\item{seed_}{Random seed.}
}
\value{
A list with the following elements
  \describe{
    \item{\code{nmf_result}}{The \code{NMFfitX1} object returned by the \code{nmf} call.}
    \item{\code{gene_info}}{A \code{data_frame} detailing the bases, weighted bases, 
       and D-score for each gene. The data frame is sorted by their D-scores}
    \item{\code{d_score_frequency_plot}}{Kernal distribution for the D-scores, separated by the coef components they represent.}
    \item{\code{ordered_sample_ids}}{The pseudo-order of samples calculated by sorting the differences of coef values for each sample.} 
    \item{\code{coef_line_dat}}{Plotting data for \code{coef_line_plot}}
    \item{\code{coef_line_plot}}{A line plot showing the trend formed by the coef values, using the order in \code{ordered_sample_ids}.}
    \item{\code{path_dat}}{Plotting data for \code{pca_path_plot} and \code{mds_path_plot}.}
    \item{\code{pca_path_plot}}{Connecting the PCA plot of the expression matrix using the order in \code{ordered_sample_ids}.
       The thickness of the lines indicate the jump of differences of the coef values.} 
    \item{\code{mds_path_plot}}{Same as \code{pca_path_plot}, except for MDS plot. The distance is "1 - Pearson Correlation".}
    \item{\code{top_genes_dat}}{Plotting data for \code{top_genes_free_y_plot} and \code{top_genes_fixed}.}
    \item{\code{top_genes_free_y_plot}}{Faceted bar-plot showing the expression levels of the top 100 genes, across all samples.
       Samples are sorted using the order in \code{ordered_sample_ids}.}
    \item{\code{top_genes_fixed}}{Sample as \code{top_genes_free_y_plot}, except all y-axes are fixed across all faceted.}
  }
}
\description{
Detect subpopulations using NMF
}
\examples{
res <- nmf_subpopulation(toy)
}

