% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phe}
\alias{phe}
\title{Grouping vector of distal mouse lung epithelial cells from two developmental stages (E14.5 and E16.5)}
\format{A vector indicating the E14.5 and E16.5 samples.}
\usage{
phe

phe
}
\description{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52583

Phenotype vector for distal mouse lung epithelial cells dataset
}
\keyword{datasets}

