% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_generation.R
\name{plot_module_graph}
\alias{plot_module_graph}
\title{Plot module graph}
\usage{
plot_module_graph(graphs_, ppi_, top_modules_, from_, to_, seed_ = 12345,
  verbose_level_ = 1)
}
\description{
Plot module graph
}

