% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Q}
\alias{Q}
\title{Log-likelihood function from M-step, when t is estimated}
\usage{
Q(Xtt, Theta, Lambda)
}
\arguments{
\item{Xtt}{matrix. Correspond to the matrix multiplication of t(X) and t.}

\item{Theta}{matrix of dimension \code{M x H}.}

\item{Lambda}{matrix of dimension \code{H x K}.}
}
\value{
Returns the log-likelihood function from M-step, when t is estimated.
}
\description{
Log-likelihood function from M-step, when t is estimated
}
\keyword{internal}
