% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_generation.R
\name{spinglass_procedure}
\alias{spinglass_procedure}
\title{Generate modules from seed genes using Spinglass community detection method}
\usage{
spinglass_procedure(expr_matrix_, grouping_vec_, selected_genes_, ppi_edges_,
  organism_, log_transformation_ = T, verbose_level_ = 1,
  n_top_modules_ = 5, n_threads_ = 1, min_size_ = 10, max_size_ = 100,
  enrichment_analysis_ = "GO", gamma_ = 0.5, seed_ = 12345, ...)
}
\arguments{
\item{expr_matrix_}{The expression matrix, each row is a gene and each
column is a sample. The row names should be gene symbols and the column
names should be sample ids. Use the correct capticalization for gene
symbols. For mouse genes,only the first letter is capitalized (e.g. Tp53);
for human genes, all letters are capitalized (e.g. TP53). The values will
be log-transformed by default, this can be changed using the
\code{log_transofrmation_} option.}

\item{grouping_vec_}{A factor vector, indicating the groupping. Its length
should be the same as the number of columns in \code{expr_matrix_}. Example:
\code{factor(c(1,1,1,2,2,2,2,2,2))}}

\item{selected_genes_}{A character vector. The genes that are used as seed
genes in the FEM algorithm. Typically generated from another method such as
NMF.}

\item{ppi_edges_}{A \code{data.frame} containing all the protein-protein interactions (PPI).
It should have two columns of gene symbols, each row indicating a pair of genes that have interaction.
\code{data(hppi)} and \code{data(mppi)} are two example PPI networks, respectively for humans and mice.}

\item{organism_}{Either "human" or "mouse".}

\item{log_transformation_}{Whether to log-transform the expression matrix.
A pseudo count \code{1} will be added to each value before log transformation
to avoid infinity. That is, \code{expr_matrix_ <- log(expr_matrix_+1)}. Default: \code{True}.}

\item{verbose_level_}{How much information is printed. Either 0, 1 or 2,
default: 1.}

\item{n_top_modules_}{Number of top modules to pick. Default: 5.}

\item{n_threads_}{Number of threads to use. Default: 1.}

\item{min_size_}{Module size lower cut-off. Default: 10.}

\item{max_size_}{Module size upper cut-off. Default: 100.}

\item{enrichment_analysis_}{Either "GO" or "KEGG".}

\item{seed_}{Random seed.}

\item{...}{Additional parameters for \code{KEGG_analysis} or \code{GO_analysis}.}
}
\description{
Generate modules from seed genes using Spinglass community detection method
}

