% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{t_update}
\alias{t_update}
\title{Update of t}
\usage{
t_update(X, Theta, Lambda, p)
}
\arguments{
\item{X}{a matrix of dimension \code{N} (number of observation) \code{x M} (number of variables) containing multinomials observations.}

\item{Theta}{matrix of dimension \code{M x H}.}

\item{Lambda}{matrix of dimension \code{H x K}.}

\item{p}{vector containing the proportions of each cluster. Must be of dimension \code{K}.}
}
\value{
Returns the updated value of Theta
}
\description{
Update of t
}
\keyword{internal}
