% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{theta_update}
\alias{theta_update}
\title{Update of theta}
\usage{
theta_update(Xtt, Theta, Lambda)
}
\arguments{
\item{Xtt}{matrix. Correspond to the matrix multiplication of t(X) and t.}

\item{Theta}{matrix of dimension \code{M x H}.}

\item{Lambda}{matrix of dimension \code{H x K}.}
}
\value{
Returns the updated value of Theta
}
\description{
Update of theta
}
\keyword{internal}
