% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{travelers}
\alias{travelers}
\title{Time profiles of 47,000 travelers.}
\format{A data frame with 47,757 rows and 166 columns.
\describe{
  \item{ID_card}{smart card encrypted identifier}
  \item{Friday00}{number of travels made on Fridays between 12:00 a.m and 12:59 a.m}
  \item{Friday01}{number of travels made on Fridays between 1:00 a.m and 1:59 a.m}
  \item{Friday02}{number of travels made on Fridays between 2:00 a.m and 2:59 a.m}
  \item{Friday03}{number of travels made on Fridays between 3:00 a.m and 3:59 a.m}
  \item{Friday04}{number of travels made on Fridays between 4:00 a.m and 4:59 a.m}
  \item{Friday05}{number of travels made on Fridays between 5:00 a.m and 5:59 a.m}
  \item{Friday06}{number of travels made on Fridays between 6:00 a.m and 6:59 a.m}
  \item{Friday07}{number of travels made on Fridays between 7:00 a.m and 7:59 a.m}
  \item{Friday08}{number of travels made on Fridays between 8:00 a.m and 8:59 a.m}
  \item{Friday09}{number of travels made on Fridays between 9:00 a.m and 9:59 a.m}
  \item{Friday10}{number of travels made on Fridays between 10:00 a.m and 10:59 a.m}
  \item{Friday11}{number of travels made on Fridays between 11:00 a.m and 11:59 a.m}
  \item{Friday12}{number of travels made on Fridays between 12:00 p.m and 12:59 p.m}
  \item{Friday13}{number of travels made on Fridays between 1:00 p.m and 1:59 p.m}
  \item{Friday14}{number of travels made on Fridays between 2:00 p.m and 2:59 p.m}
  \item{Friday15}{number of travels made on Fridays between 3:00 p.m and 3:59 p.m}
  \item{Friday16}{number of travels made on Fridays between 4:00 p.m and 4:59 p.m}
  \item{Friday17}{number of travels made on Fridays between 5:00 p.m and 5:59 p.m}
  \item{Friday18}{number of travels made on Fridays between 6:00 p.m and 6:59 p.m}
  \item{Friday19}{number of travels made on Fridays between 7:00 p.m and 7:59 p.m}
  \item{Friday20}{number of travels made on Fridays between 8:00 p.m and 8:59 p.m}
  \item{Friday21}{number of travels made on Fridays between 9:00 p.m and 9:59 p.m}
  \item{Friday22}{number of travels made on Fridays between 10:00 p.m and 10:59 p.m}
  \item{Friday23}{number of travels made on Fridays between 11:00 p.m and 11:59 p.m}
  \item{Monday00}{number of travels made on Mondays between 12:00 a.m and 12:59 a.m}
  \item{Monday01}{number of travels made on Mondays between 1:00 a.m and 1:59 a.m}
  \item{Monday03}{number of travels made on Mondays between 3:00 a.m and 3:59 a.m}
  \item{Monday04}{number of travels made on Mondays between 4:00 a.m and 4:59 a.m}
  \item{Monday05}{number of travels made on Mondays between 5:00 a.m and 5:59 a.m}
  \item{Monday06}{number of travels made on Mondays between 6:00 a.m and 6:59 a.m}
  \item{Monday07}{number of travels made on Mondays between 7:00 a.m and 7:59 a.m}
  \item{Monday08}{number of travels made on Mondays between 8:00 a.m and 8:59 a.m}
  \item{Monday09}{number of travels made on Mondays between 9:00 a.m and 9:59 a.m}
  \item{Monday10}{number of travels made on Mondays between 10:00 a.m and 10:59 a.m}
  \item{Monday11}{number of travels made on Mondays between 11:00 a.m and 11:59 a.m}
  \item{Monday12}{number of travels made on Mondays between 12:00 p.m and 12:59 p.m}
  \item{Monday13}{number of travels made on Mondays between 1:00 p.m and 1:59 p.m}
  \item{Monday14}{number of travels made on Mondays between 2:00 p.m and 2:59 p.m}
  \item{Monday15}{number of travels made on Mondays between 3:00 p.m and 3:59 p.m}
  \item{Monday16}{number of travels made on Mondays between 4:00 p.m and 4:59 p.m}
  \item{Monday17}{number of travels made on Mondays between 5:00 p.m and 5:59 p.m}
  \item{Monday18}{number of travels made on Mondays between 6:00 p.m and 6:59 p.m}
  \item{Monday19}{number of travels made on Mondays between 7:00 p.m and 7:59 p.m}
  \item{Monday20}{number of travels made on Mondays between 8:00 p.m and 8:59 p.m}
  \item{Monday21}{number of travels made on Mondays between 9:00 p.m and 9:59 p.m}
  \item{Monday22}{number of travels made on Mondays between 10:00 p.m and 10:59 p.m}
  \item{Monday23}{number of travels made on Mondays between 11:00 p.m and 11:59 p.m}
  \item{Saturday00}{number of travels made on Saturdays between 12:00 a.m and 12:59 a.m}
  \item{Saturday01}{number of travels made on Saturdays between 1:00 a.m and 1:59 a.m}
  \item{Saturday02}{number of travels made on Saturdays between 2:00 a.m and 2:59 a.m}
  \item{Saturday03}{number of travels made on Saturdays between 3:00 a.m and 3:59 a.m}
  \item{Saturday04}{number of travels made on Saturdays between 4:00 a.m and 4:59 a.m}
  \item{Saturday05}{number of travels made on Saturdays between 5:00 a.m and 5:59 a.m}
  \item{Saturday06}{number of travels made on Saturdays between 6:00 a.m and 6:59 a.m}
  \item{Saturday07}{number of travels made on Saturdays between 7:00 a.m and 7:59 a.m}
  \item{Saturday08}{number of travels made on Saturdays between 8:00 a.m and 8:59 a.m}
  \item{Saturday09}{number of travels made on Saturdays between 9:00 a.m and 9:59 a.m}
  \item{Saturday10}{number of travels made on Saturdays between 10:00 a.m and 10:59 a.m}
  \item{Saturday11}{number of travels made on Saturdays between 11:00 a.m and 11:59 a.m}
  \item{Saturday12}{number of travels made on Saturdays between 12:00 p.m and 12:59 p.m}
  \item{Saturday13}{number of travels made on Saturdays between 1:00 p.m and 1:59 p.m}
  \item{Saturday14}{number of travels made on Saturdays between 2:00 p.m and 2:59 p.m}
  \item{Saturday15}{number of travels made on Saturdays between 3:00 p.m and 3:59 p.m}
  \item{Saturday16}{number of travels made on Saturdays between 4:00 p.m and 4:59 p.m}
  \item{Saturday17}{number of travels made on Saturdays between 5:00 p.m and 5:59 p.m}
  \item{Saturday18}{number of travels made on Saturdays between 6:00 p.m and 6:59 p.m}
  \item{Saturday19}{number of travels made on Saturdays between 7:00 p.m and 7:59 p.m}
  \item{Saturday20}{number of travels made on Saturdays between 8:00 p.m and 8:59 p.m}
  \item{Saturday21}{number of travels made on Saturdays between 9:00 p.m and 9:59 p.m}
  \item{Saturday22}{number of travels made on Saturdays between 10:00 p.m and 10:59 p.m}
  \item{Saturday23}{number of travels made on Saturdays between 11:00 p.m and 11:59 p.m}
  \item{Sunday00}{number of travels made on Sundays between 12:00 a.m and 12:59 a.m}
  \item{Sunday01}{number of travels made on Sundays between 1:00 a.m and 1:59 a.m}
  \item{Sunday02}{number of travels made on Sundays between 2:00 a.m and 2:59 a.m}
  \item{Sunday03}{number of travels made on Sundays between 3:00 a.m and 3:59 a.m}
  \item{Sunday04}{number of travels made on Sundays between 4:00 a.m and 4:59 a.m}
  \item{Sunday05}{number of travels made on Sundays between 5:00 a.m and 5:59 a.m}
  \item{Sunday06}{number of travels made on Sundays between 6:00 a.m and 6:59 a.m}
  \item{Sunday07}{number of travels made on Sundays between 7:00 a.m and 7:59 a.m}
  \item{Sunday08}{number of travels made on Sundays between 8:00 a.m and 8:59 a.m}
  \item{Sunday09}{number of travels made on Sundays between 9:00 a.m and 9:59 a.m}
  \item{Sunday10}{number of travels made on Sundays between 10:00 a.m and 10:59 a.m}
  \item{Sunday11}{number of travels made on Sundays between 11:00 a.m and 11:59 a.m}
  \item{Sunday12}{number of travels made on Sundays between 12:00 p.m and 12:59 p.m}
  \item{Sunday13}{number of travels made on Sundays between 1:00 p.m and 1:59 p.m}
  \item{Sunday14}{number of travels made on Sundays between 2:00 p.m and 2:59 p.m}
  \item{Sunday15}{number of travels made on Sundays between 3:00 p.m and 3:59 p.m}
  \item{Sunday16}{number of travels made on Sundays between 4:00 p.m and 4:59 p.m}
  \item{Sunday17}{number of travels made on Sundays between 5:00 p.m and 5:59 p.m}
  \item{Sunday18}{number of travels made on Sundays between 6:00 p.m and 6:59 p.m}
  \item{Sunday19}{number of travels made on Sundays between 7:00 p.m and 7:59 p.m}
  \item{Sunday20}{number of travels made on Sundays between 8:00 p.m and 8:59 p.m}
  \item{Sunday21}{number of travels made on Sundays between 9:00 p.m and 9:59 p.m}
  \item{Sunday22}{number of travels made on Sundays between 10:00 p.m and 10:59 p.m}
  \item{Sunday23}{number of travels made on Sundays between 11:00 p.m and 11:59 p.m}
  \item{Thursday00}{number of travels made on Thursdays between 12:00 a.m and 12:59 a.m}
  \item{Thursday01}{number of travels made on Thursdays between 1:00 a.m and 1:59 a.m}
  \item{Thursday02}{number of travels made on Thursdays between 2:00 a.m and 2:59 a.m}
  \item{Thursday03}{number of travels made on Thursdays between 3:00 a.m and 3:59 a.m}
  \item{Thursday04}{number of travels made on Thursdays between 4:00 a.m and 4:59 a.m}
  \item{Thursday05}{number of travels made on Thursdays between 5:00 a.m and 5:59 a.m}
  \item{Thursday06}{number of travels made on Thursdays between 6:00 a.m and 6:59 a.m}
  \item{Thursday07}{number of travels made on Thursdays between 7:00 a.m and 7:59 a.m}
  \item{Thursday08}{number of travels made on Thursdays between 8:00 a.m and 8:59 a.m}
  \item{Thursday09}{number of travels made on Thursdays between 9:00 a.m and 9:59 a.m}
  \item{Thursday10}{number of travels made on Thursdays between 10:00 a.m and 10:59 a.m}
  \item{Thursday11}{number of travels made on Thursdays between 11:00 a.m and 11:59 a.m}
  \item{Thursday12}{number of travels made on Thursdays between 12:00 p.m and 12:59 p.m}
  \item{Thursday13}{number of travels made on Thursdays between 1:00 p.m and 1:59 p.m}
  \item{Thursday14}{number of travels made on Thursdays between 2:00 p.m and 2:59 p.m}
  \item{Thursday15}{number of travels made on Thursdays between 3:00 p.m and 3:59 p.m}
  \item{Thursday16}{number of travels made on Thursdays between 4:00 p.m and 4:59 p.m}
  \item{Thursday17}{number of travels made on Thursdays between 5:00 p.m and 5:59 p.m}
  \item{Thursday18}{number of travels made on Thursdays between 6:00 p.m and 6:59 p.m}
  \item{Thursday19}{number of travels made on Thursdays between 7:00 p.m and 7:59 p.m}
  \item{Thursday20}{number of travels made on Thursdays between 8:00 p.m and 8:59 p.m}
  \item{Thursday21}{number of travels made on Thursdays between 9:00 p.m and 9:59 p.m}
  \item{Thursday22}{number of travels made on Thursdays between 10:00 p.m and 10:59 p.m}
  \item{Thursday23}{number of travels made on Thursdays between 11:00 p.m and 11:59 p.m}
  \item{Tuesday00}{number of travels made on Tuesdays between 12:00 a.m and 12:59 a.m}
  \item{Tuesday01}{number of travels made on Tuesdays between 1:00 a.m and 1:59 a.m}
  \item{Tuesday03}{number of travels made on Tuesdays between 3:00 a.m and 3:59 a.m}
  \item{Tuesday04}{number of travels made on Tuesdays between 4:00 a.m and 4:59 a.m}
  \item{Tuesday05}{number of travels made on Tuesdays between 5:00 a.m and 5:59 a.m}
  \item{Tuesday06}{number of travels made on Tuesdays between 6:00 a.m and 6:59 a.m}
  \item{Tuesday07}{number of travels made on Tuesdays between 7:00 a.m and 7:59 a.m}
  \item{Tuesday08}{number of travels made on Tuesdays between 8:00 a.m and 8:59 a.m}
  \item{Tuesday09}{number of travels made on Tuesdays between 9:00 a.m and 9:59 a.m}
  \item{Tuesday10}{number of travels made on Tuesdays between 10:00 a.m and 10:59 a.m}
  \item{Tuesday11}{number of travels made on Tuesdays between 11:00 a.m and 11:59 a.m}
  \item{Tuesday12}{number of travels made on Tuesdays between 12:00 p.m and 12:59 p.m}
  \item{Tuesday13}{number of travels made on Tuesdays between 1:00 p.m and 1:59 p.m}
  \item{Tuesday14}{number of travels made on Tuesdays between 2:00 p.m and 2:59 p.m}
  \item{Tuesday15}{number of travels made on Tuesdays between 3:00 p.m and 3:59 p.m}
  \item{Tuesday16}{number of travels made on Tuesdays between 4:00 p.m and 4:59 p.m}
  \item{Tuesday17}{number of travels made on Tuesdays between 5:00 p.m and 5:59 p.m}
  \item{Tuesday18}{number of travels made on Tuesdays between 6:00 p.m and 6:59 p.m}
  \item{Tuesday19}{number of travels made on Tuesdays between 7:00 p.m and 7:59 p.m}
  \item{Tuesday20}{number of travels made on Tuesdays between 8:00 p.m and 8:59 p.m}
  \item{Tuesday21}{number of travels made on Tuesdays between 9:00 p.m and 9:59 p.m}
  \item{Tuesday22}{number of travels made on Tuesdays between 10:00 p.m and 10:59 p.m}
  \item{Tuesday23}{number of travels made on Tuesdays between 11:00 p.m and 11:59 p.m}
  \item{Wednesday00}{number of travels made on Wednesdays between 12:00 a.m and 12:59 a.m}
  \item{Wednesday01}{number of travels made on Wednesdays between 1:00 a.m and 1:59 a.m}
  \item{Wednesday03}{number of travels made on Wednesdays between 3:00 a.m and 3:59 a.m}
  \item{Wednesday04}{number of travels made on Wednesdays between 4:00 a.m and 4:59 a.m}
  \item{Wednesday05}{number of travels made on Wednesdays between 5:00 a.m and 5:59 a.m}
  \item{Wednesday06}{number of travels made on Wednesdays between 6:00 a.m and 6:59 a.m}
  \item{Wednesday07}{number of travels made on Wednesdays between 7:00 a.m and 7:59 a.m}
  \item{Wednesday08}{number of travels made on Wednesdays between 8:00 a.m and 8:59 a.m}
  \item{Wednesday09}{number of travels made on Wednesdays between 9:00 a.m and 9:59 a.m}
  \item{Wednesday10}{number of travels made on Wednesdays between 10:00 a.m and 10:59 a.m}
  \item{Wednesday11}{number of travels made on Wednesdays between 11:00 a.m and 11:59 a.m}
  \item{Wednesday12}{number of travels made on Wednesdays between 12:00 p.m and 12:59 p.m}
  \item{Wednesday13}{number of travels made on Wednesdays between 1:00 p.m and 1:59 p.m}
  \item{Wednesday14}{number of travels made on Wednesdays between 2:00 p.m and 2:59 p.m}
  \item{Wednesday15}{number of travels made on Wednesdays between 3:00 p.m and 3:59 p.m}
  \item{Wednesday16}{number of travels made on Wednesdays between 4:00 p.m and 4:59 p.m}
  \item{Wednesday17}{number of travels made on Wednesdays between 5:00 p.m and 5:59 p.m}
  \item{Wednesday18}{number of travels made on Wednesdays between 6:00 p.m and 6:59 p.m}
  \item{Wednesday19}{number of travels made on Wednesdays between 7:00 p.m and 7:59 p.m}
  \item{Wednesday20}{number of travels made on Wednesdays between 8:00 p.m and 8:59 p.m}
  \item{Wednesday21}{number of travels made on Wednesdays between 9:00 p.m and 9:59 p.m}
  \item{Wednesday22}{number of travels made on Wednesdays between 10:00 p.m and 10:59 p.m}
  \item{Wednesday23}{number of travels made on Wednesdays between 11:00 p.m and 11:59 p.m}
}}
\source{
These data are aggregated from Transdev datasets.
}
\usage{
travelers
}
\description{
A dataset containing the number of travels made at every moment of the week
during one month on a public transportation network for more than 47,000 travelers.
}
\keyword{datasets}
