% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phe}
\alias{phe}
\title{Phenotype vector for distal mouse lung epithelial cells dataset}
\format{A data frame with five variables: \code{year}, \code{sex},
  \code{name}, \code{n} and \code{prop} (\code{n} divided by total number
  of applicants in that year).}
\usage{
phe
}
\description{
Phenotype vector for distal mouse lung epithelial cells dataset
}
\keyword{datasets}

